+++ packages/algob/src/lib/lsig.ts
@@ -14,8 +14,7 @@ import { CompileOp } from "./compile";
 export async function getLsig (
   name: string,
   algodClient: Algodv2,
-  scTmplParams?: SCParams):
-  Promise<LogicSigAccount> {
+  scTmplParams?: SCParams): Promise<LogicSigAccount> {
   const compileOp = new CompileOp(algodClient);
   const result: ASCCache = await compileOp.ensureCompiled(name, false, scTmplParams);
   const program = result.base64ToBytes;