+++ CHANGELOG.md
@@ -9,9 +9,16 @@
     * Extract, Extract3 opcodes
     * ExtractUint16, ExtractUint32, ExtractUint64 opcodes
     * Txn, Global fields
+    * Added application account (a smart contract now has an escrow account). Updated checkpoint structure to store `applicationAccount` while running `algob` scripts.
+
+### Breaking changes
+
+`@algo-builder/runtime`:
++ Renamed `Runtime.getLogicSig` to `Runtime.createLsigAccount` #506.
 
 ### Bug Fixes
 +  Fix bug substring3 opcode pop wrong order [/#505](https://github.com/scale-it/algo-builder/pull/505), contribution: @vuvth.
++ Fix bug: `runtime.optinToApp` updating state even after opt-in fails. Reported by @patrick
 
 ## v2.1.0 2021-10-22